+++ cni/network/network_linux.go
@@ -1,0 +1,11 @@
+package network
+
+import (
+	"github.com/Azure/azure-container-networking/cni"
+	cniTypesCurr "github.com/containernetworking/cni/pkg/types/current"
+)
+
+// HandleConsecutiveAdd is a dummy function for Linux platform.
+func HandleConsecutiveAdd(containerId, endpointId string, dnsServers []string, nwInfo *NetworkInfo, nwCfg *cni.NetworkConfig) (*cniTypesCurr.Result, error) {
+	return nil, nil
+}
