+++ network/network.go
@@ -6,6 +6,7 @@ package network
 import (
 	"net"
 
+	"github.com/Azure/azure-container-networking/cni"
 	"github.com/Azure/azure-container-networking/log"
 	"github.com/Azure/azure-container-networking/platform"
 )
@@ -46,7 +47,7 @@ type NetworkInfo struct {
 	Mode       string
 	Subnets    []SubnetInfo
 	DNS        DNSInfo
-	Policies   []Policy
+	Policies   []cni.Policy
 	BridgeName string
 	Options    map[string]interface{}
 }
